sin discovery

+++ tarjetacredito/src/main/java/edu/um/ar/programacion2/tarjetacredito/objeto/TarjetaCreditoObjeto.java
@@ -8,7 +8,7 @@ public class TarjetaCreditoObjeto {
 
 	private Long id;
 	private String tipo;
-	private Integer numero;
+	private Long numero;
 	private Integer codseguridad;
 	private Date vencimiento;
 	private Float montomaximo;
@@ -32,11 +32,11 @@ public class TarjetaCreditoObjeto {
 		this.tipo = tipo;
 	}
 
-	public Integer getNumero() {
+	public Long getNumero() {
 		return numero;
 	}
 
-	public void setNumero(Integer numero) {
+	public void setNumero(Long numero) {
 		this.numero = numero;
 	}
 
@@ -85,7 +85,7 @@ public class TarjetaCreditoObjeto {
 		this.activo = activo;
 	}
 
-	public TarjetaCreditoObjeto(Long id, String tipo, Integer numero, Integer codseguridad, Date vencimiento,
+	public TarjetaCreditoObjeto(Long id, String tipo, Long numero, Integer codseguridad, Date vencimiento,
 			Float montomaximo, Long cliente_id, boolean activo) {
 		super();
 		this.id = id;
