+++ tarjetacredito/src/main/java/edu/um/ar/programacion2/tarjetacredito/repository/TarjetaCreditoRepository.java
@@ -9,7 +9,7 @@ import edu.um.ar.programacion2.tarjetacredito.objeto.TarjetaCreditoObjeto;
 
 public interface TarjetaCreditoRepository extends JpaRepository<TarjetaCredito, Long> {
 	Optional<TarjetaCredito> findByNumero(Integer numero);
-	Boolean existsByNumero(Integer numero);
+	Boolean existsByNumero(Long numero);
 	Boolean existsByToken(String token);
 	Optional<TarjetaCredito> findByToken(String token);
 	Optional<TarjetaCredito> findById(Long id);
